loop-api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.10
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.1
openai==1.51.2
orjson==3.10.7
loguru==0.7.2
